Moved login to tabs to fix issue of rendering the right page when launching the app
Fixed the issue not being able to click on tab bar icons and added the login to the tabs directory just for easier testing when developing the app.
