+++ app/(tabs)/login.tsx
@@ -1,3 +1,4 @@
+// moved the login file to tabs for now
 import { useState } from "react";
 import { View, Text, TextInput, Button, StyleSheet } from "react-native";
